+++ CHANGELOG.md
@@ -10,6 +10,9 @@
 - Moved DELETE-SCRIPT functionality to `Ctrl`+`X` under SCRIPTS
 - Added `y` key binding to copy in LINKS
 - Changed key binding for rotation of key binding info grid from `?` to `<` and `>`
+- Changed to using left right arrows to adjust screen split
+    - Left right arrows no longer control undo/redo
+    - `Tab`/`Shift`+`Tab` no longer control screen split
 - Added "focus content" via `z` which hides the bottom right pane
 - Added key binding info pane toggle via `?`
 - Made the UX of various input fields slightly more polished